requirements.txt
python-telegram-bot[job-queue]==21.10
Pillow>=10.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
redis[hiredis]>=5.0.0,<6
